refactor(tests): unify stream capture in tests/support/streams.ts
Three test files had the same vi.spyOn-on-process.stderr-write pattern:
inspect.test.ts had a generic captureStream(stream), plans.test.ts had
an installStderrCapture() variant that also passed through to real
stderr, and wal.test.ts (added in this PR) inlined a third copy with
vi.spyOn.

Extract the inspect.test.ts shape (captures + suppresses, returning
{chunks, restore}) into tests/support/streams.ts and route all three
through it. The plans.test.ts variant's pass-through behavior was
incidental — its only assertion is on chunks.join(), so suppression
keeps test output cleaner without changing what's tested.
